2.0：dm(data manager)模块完成,单元测试通过

+++ src/main/java/top/wuzonghui/simpledb/backend/dm/page/PageX.java
@@ -79,7 +79,7 @@ public class PageX {
      * @return
      */
     public static int getFreeSpace(Page pg) {
-        return PageCache.PAGE_SIZE - getFSO(pg);
+        return PageCache.PAGE_SIZE - (int) getFSO(pg);
     }
 
     /**
